// FlipBook/src/app/data/interface/cuadernillo.interface.ts
export interface Cuadernillo {
  expCuardenillo: string;
  area: string;
  folio: string;
  fecha: string;
  nombreDenunciante: string;
  appDenunciante: string;
  apmDenunciante: string;
  nombreAcusado: string;
  nombreJuez:string;
  cargoJuez:string;
  nombreSubdirector: string;
  cargoSubdirector: string;

}